+++ R/dq_icon.R
@@ -18,7 +18,7 @@
 #' if (interactive()) {
 #'
 #' library(shiny)
-#' addResourcePath("images", system.file("www", "img", package = "dqutils"))
+#' addResourcePath("images", system.file("www", "img", package = "dqshiny"))
 #' shinyApp(
 #'   ui = fluidPage(
 #'     dq_icon("table", size = "4x"),
@@ -32,7 +32,7 @@
 #' }
 dq_icon <- function(
   icon, lib = "font-awesome",
-  color = "#ff8f00", bg_color = "#fff", size = NULL, ...
+  color = "#ff8f00", bg_color = NULL, size = NULL, ...
 ) {
   if (length(size) > 0 && !(size %in% c("xs", "sm", "lg", paste0(2:10, "x")))) {
     size <- NULL