+++ R/hot_renderer.R
@@ -34,10 +34,11 @@ dq_hot_date_renderer <- function() {
 #' @export
 #' @rdname dq_hot_date_renderer
 dq_hot_empty_renderer <- function(renderer = "Autocomplete") {
-  if (length(renderer) == 0 ||
-      !(renderer %in% c("Autocomplete", "Base", "Checkbox", "Date", "Dropdown",
-                        "Html", "Numeric", "Password", "Text", "Time")))
+  if (length(renderer) != 1L || !(renderer %in% c(
+    "Autocomplete", "Base", "Checkbox", "Date", "Dropdown",
+    "Html", "Numeric", "Password", "Text", "Time"))) {
     renderer <- "Autocomplete"
+  }
   paste0("function (instance, td, row, col, prop, value, cellProperties) {
     Handsontable.renderers.", renderer, "Renderer.apply(this, arguments);
     if (value.toString() == ' ' || value.toString() == '') {
@@ -56,8 +57,11 @@ dq_hot_html_renderer <- function() {
    }"
 }
 
-#' TODO use for dq_add_selectize_options
-selectizeRenderer <- function() {
+#' @description dq_hot_selectize_renderer: Renderer to properly display multiple
+#' selectize options.
+#' @export
+#' @rdname dq_hot_date_renderer
+dq_hot_selectize_renderer <- function() {
   "function (instance, td, row, col, prop, value, cellProperties) {
     var settings = cellProperties.selectizeOptions;
     if (value && settings && (settings.maxItems === null || settings.maxItems > 1)) {