+++ tests/testthat/test-js.R
@@ -2,9 +2,9 @@ context("js / add_js")
 
 test_that("add_js works with any input", {
   if (requireNamespace("V8", quietly = TRUE)) {
-    expect_error(add_js(NULL, NULL), "SyntaxError")
+    expect_error(add_js(NULL, NULL), "Error")
     expect_error(add_js(NULL, "function_text"), "ReferenceError")
-    expect_error(add_js("type", NULL), "SyntaxError")
+    expect_error(add_js("type", NULL), "Error")
     expect_silent(add_js("type", "function(params) {
        alert('Result is: ' + (parseInt(params[0]) + parseInt(params[1])));
      }"))
@@ -21,11 +21,20 @@ test_that("add_js works with any input", {
 context("js / run_js")
 
 test_that("run_js works with any input", {
-  expect_null(run_js(NULL, NULL))
-  expect_null(run_js(NULL, "params"))
-  expect_null(run_js("type", NULL))
-  expect_null(run_js("type", c("my", "params", "are")))
-  expect_null(run_js("type", list("my", "params", "are")))
+  session <- create_test_session("", NULL, NULL)
+  shiny::withReactiveDomain(session, {
+    expect_silent(run_js(NULL, NULL))
+    expect_silent(run_js(NULL, "params"))
+    expect_silent(run_js("type", NULL))
+    expect_silent(run_js("type", c("my", "params", "are")))
+    expect_silent(run_js("type", list("my", "params", "are")))
+  })
+  m <- session$lastCustomMessages
+  expect_null(m[[1]]$type)
+  expect_equal(m[[2]]$message[[1]], "params")
+  expect_null(m[[3]]$message[[1]])
+  expect_length(m[[4]]$message[[1]], 3)
+  expect_length(m[[5]]$message[[1]], 3)
 })
 
 context("js / parse_code")