+++ NEWS.md
@@ -1,3 +1,13 @@
+# dqshiny 0.0.2
+
+* Fix text filter to properly skip empty filters.
+* Add borders and headers to dq_htmltable.
+* Add 'base-features' demo.
+* Support reactiveVal as dq_hot data.
+* Support NA filter to detect proper filter values.
+* Fix autocomplete dropdown display for filtering.
+* Automatically update filter values when dq_hot data changes.
+
 # dqshiny 0.0.1
 
 * Initial release.